+++ vaadin-ordered-layout-flow/src/main/java/com/vaadin/flow/component/orderedlayout/FlexComponent.java
@@ -87,50 +87,6 @@ public interface FlexComponent<C extends Component>
     }
 
     /**
-     * Possible values for the {@code flex-direction} CSS property, which determines how the elements are placed
-     * inside the layout.
-     */
-    public enum FlexDirection {
-
-        /**
-         * The items are displayed horizontally, as a row.
-         */
-        ROW("row"),
-
-        /**
-         * The items are displayed horizontally, as a row in reverse order.
-         */
-        ROW_REVERSE("row-reverse"),
-
-        /**
-         * The items are displayed vertically, as a column.
-         */
-        COLUMN("column"),
-
-        /**
-         * The items are displayed vertically, as a column in reverse order.
-         */
-        COLUMN_REVERSE("column-reverse");
-
-        private final String directionValue;
-
-        FlexDirection(String directionValue) {
-            this.directionValue = directionValue;
-        }
-
-        String getDirectionValue() {
-            return directionValue;
-        }
-
-        static FlexDirection toFlexDirection(String flexValue, FlexDirection defaultValue) {
-            return Arrays.stream(values())
-                    .filter(flexDirection -> flexDirection.getDirectionValue()
-                            .equals(flexValue))
-                    .findFirst().orElse(defaultValue);
-        }
-    }
-
-    /**
      * Enum with the possible values for the way the extra space inside the
      * layout is distributed among the components. It correlates to the
      * <code>justify-content</code> CSS property.
@@ -382,172 +338,6 @@ public interface FlexComponent<C extends Component>
      */
     default public void expand(Component... componentsToExpand) {
         setFlexGrow(1.0d, componentsToExpand);
-    }
-
-    /**
-     * Sets the flex basis property of the components inside the layout.
-     * The flex basis property specifies the initial main size of a component.
-     *
-     * @param width
-     *            the width for the components
-     * @param elementContainers
-     *            the containers (components) to apply the flex basis
-     *            property. Setting <code>null</code> will remove the flex
-     *            basis property
-     */
-    default public void setFlexBasis(String width,
-                             HasElement... elementContainers) {
-        if (width == null) {
-            for (HasElement element : elementContainers) {
-                element.getElement().getStyle().remove(
-                        FlexConstants.FLEX_BASIS_CSS_PROPERTY);
-            }
-        } else {
-            for (HasElement element : elementContainers) {
-                element.getElement().getStyle().set(
-                        FlexConstants.FLEX_BASIS_CSS_PROPERTY,
-                        width);
-            }
-        }
-    }
-
-    /**
-     * Gets the flex basis property of a given element container.
-     *
-     * @param elementContainer
-     *            the element container to read the flex basis property from
-     * @return the flex grow property
-     */
-    default public String getFlexBasis(HasElement elementContainer) {
-        return elementContainer.getElement().getStyle()
-                .get(FlexConstants.FLEX_BASIS_CSS_PROPERTY);
-    }
-
-    /**
-     * Sets the flex direction property of the layout.
-     * The flex direction property specifies how components are placed in the
-     * layout defining the main axis and the direction (normal or reversed).
-     *
-     * @param flexDirection
-     *            the direction for the components
-     */
-    default public void setFlexDirection(FlexDirection flexDirection) {
-        if (flexDirection == null) {
-            getElement().getStyle().remove(
-                    FlexConstants.FLEX_DIRECTION_CSS_PROPERTY);
-        } else {
-            getElement().getStyle().set(
-                    FlexConstants.FLEX_DIRECTION_CSS_PROPERTY,
-                    flexDirection.getDirectionValue());
-        }
-    }
-
-    /**
-     * Gets the flex direction property of a given element container.
-     *
-     * @param elementContainer
-     *            the element container to read the flex direction property from
-     * @return the flex direction property,
-     *         or {@link FlexDirection#ROW} if none was set
-     */
-    default public FlexDirection getFlexDirection(HasElement elementContainer) {
-        return FlexDirection.toFlexDirection(elementContainer.getElement().getStyle()
-                        .get(FlexConstants.FLEX_DIRECTION_CSS_PROPERTY),
-                FlexDirection.ROW);
-    }
-
-    /**
-     * Sets the flex shrink property of the components inside the layout. The flex
-     * shrink property specifies how the item will shrink relative to the rest
-     * of the components inside the same layout.
-     *
-     * Negative values are not allowed.
-     *
-     * @param flexShrink
-     *            how much the component will shrink relative to the rest
-     *            of the components
-     * @param elementContainers
-     *            the containers (components) to apply the flex shrink property
-     */
-    default public void setFlexShrink(double flexShrink,
-                              HasElement... elementContainers) {
-        if (flexShrink < 0) {
-            throw new IllegalArgumentException(
-                    "Flex shrink property cannot be negative");
-        }
-
-        for (HasElement container : elementContainers) {
-            container.getElement().getStyle().set(
-                    FlexConstants.FLEX_SHRINK_CSS_PROPERTY,
-                    String.valueOf(flexShrink));
-        }
-    }
-
-    /**
-     * Gets the flex shrink property of a given element container.
-     *
-     * @param elementContainer
-     *            the element container to read the flex shrink property from
-     * @return the flex shrink property
-     */
-    default public double getFlexShrink(HasElement elementContainer) {
-        String ratio = elementContainer.getElement().getStyle()
-                .get(FlexConstants.FLEX_SHRINK_CSS_PROPERTY);
-        if (ratio == null || ratio.isEmpty()) {
-            return 1;
-        }
-
-        try {
-            return Double.parseDouble(ratio);
-        } catch (Exception e) {
-            throw new IllegalStateException(
-                    "The flex shrink property of the element container is not parseable to double: " + ratio, e);
-        }
-    }
-
-    /**
-     * Sets the order property of the component inside the layout. The order
-     * property specifies the order of a component relative to the
-     * rest of the components inside the same layout.
-     *
-     * Setting the order property value to 0 removes the order of the element container.
-     *
-     * @param order
-     *            the order for the component
-     * @param elementContainer
-     *            the container (component) to apply the order property
-     */
-    default public void setOrder(int order, HasElement elementContainer) {
-        if (order == 0) {
-            elementContainer.getElement().getStyle().remove(
-                    FlexConstants.ORDER_CSS_PROPERTY);
-        }
-        elementContainer.getElement().getStyle().set(
-                FlexConstants.ORDER_CSS_PROPERTY,
-                String.valueOf(order));
-    }
-
-    /**
-     * Gets the order property of a given element container.
-     *
-     * @param elementContainer
-     *            the element container to read the order property from
-     * @return the order property,
-     *         or 0 if none was set
-     */
-    default public int getOrder(HasElement elementContainer) {
-        String order = elementContainer.getElement().getStyle()
-                .get(FlexConstants.ORDER_CSS_PROPERTY);
-
-        if (order == null || order.isEmpty()) {
-            return 0;
-        }
-        try {
-            return Integer.parseInt(order);
-        } catch (Exception e) {
-            throw new IllegalStateException(
-                    "The order property of the element container is not parseable to integer: " + order, e);
-        }
     }
 
     @Override
